[ADD] sector to family [IMP] operational sector and area views [MOV] 'health center' menu id from his to medical [FIX] area.sector_id (m2o) => area.sector_ids (o2m) [CHG] constraints from py to sql + add method to duplicate documents [ADD] acls
