+++ medical_his/views/medical_hospital_menu.xml
@@ -2,29 +2,28 @@
 <openerp>
     <data>
 
-        <menuitem id="medical_health_center_root" name="Health Center"
-            parent="medical.medical_root_sub" sequence="10" />
+        <menuitem id="medical.medical_health_center_root" name="Health Center"
+            parent="medical.medical_root_sub" sequence="50" />
 
-        <menuitem id="medical_hospital_bed" name=" Hospital Bed"
-            parent="medical_health_center_root" sequence="10"
+        <menuitem id="medical_hospital_bed"
+            parent="medical.medical_health_center_root" sequence="10"
             action="medical_hospital_bed_action" />
 
-        <menuitem id="medical_hospital_room" name=" Hospital Room"
-            parent="medical_health_center_root" sequence="20"
+        <menuitem id="medical_hospital_room"
+            parent="medical.medical_health_center_root" sequence="20"
             action="medical_hospital_room_action" />
 
-        <menuitem id="medical_hospital_zone" name=" Hospital Zone"
-            parent="medical_health_center_root" sequence="30"
+        <menuitem id="medical_hospital_zone"
+            parent="medical.medical_health_center_root" sequence="30"
             action="medical_hospital_zone_action" />
 
-        <menuitem id="medical_hospital_or" name=" Hospital Or"
-            parent="medical_health_center_root" sequence="40"
+        <menuitem id="medical_hospital_or"
+            parent="medical.medical_health_center_root" sequence="40"
             action="medical_hospital_or_action" />
 
-        <menuitem id="medical_hospital_unit" name=" Hospital Unit"
-            parent="medical_health_center_root" sequence="50"
+        <menuitem id="medical_hospital_unit"
+            parent="medical.medical_health_center_root" sequence="50"
             action="medical_hospital_unit_action" />
-
 
     </data>
 </openerp>
