+++ medical_operational/models/__init__.py
@@ -2,3 +2,4 @@
 
 from . import medical_operational_area
 from . import medical_operational_sector
+from . import medical_family
