+++ medical_operational/views/medical_operational_menu.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<openerp>
+    <data>
+
+        <menuitem id="medical.medical_health_center_root" name="Health Center"
+            parent="medical.medical_root_sub" sequence="50" />
+
+        <menuitem id="medical_operational_area"
+            parent="medical.medical_health_center_root"
+            action="action_medical_operational_area_view" sequence="60" />
+
+        <menuitem id="medical_operational_sector"
+            parent="medical.medical_health_center_root"
+            action="action_medical_operational_sector_view" sequence="70" />
+
+    </data>
+</openerp>
