+++ medical_operational/views/medical_operational_sector_view.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <openerp>
     <data>
+
         <record id="view_medical_operational_sector_tree" model="ir.ui.view">
             <field name="name">medical.operational_sector.tree</field>
             <field name="model">medical.operational.sector</field>
-            <field name="priority" eval="8" />
             <field name="arch" type="xml">
-                <tree string="Oemedical Operational_Sector">
+                <tree string="Operational Sectors">
                     <field name="name" />
                     <field name="area_id" />
                 </tree>
@@ -16,12 +16,24 @@
         <record id="view_medical_operational_sector_form" model="ir.ui.view">
             <field name="name">medical.operational.sector.form</field>
             <field name="model">medical.operational.sector</field>
-            <field name="priority" eval="8" />
             <field name="arch" type="xml">
-                <form string="Oemedical Operational_Sector">
-                    <field name="name" />
-                    <field name="area_id" />
-                    <field name="notes" colspan="4" nolabel="1" />
+                <form string="Operational Sector">
+                    <header></header>
+                    <sheet>
+                        <div class="oe_title" name="title">
+                            <label for="name" class="oe_edit_only"/>
+                            <h1><field name="name" /></h1>
+                            <label for="area_id" class="oe_edit_only"/>
+                            <h1><field name="area_id" /></h1>
+                        </div>
+                        <div class="oe_right oe_button_box" name="buttons"></div>
+
+                        <notebook>
+                            <page string="Notes" name="notes">
+                                <field name="notes"/>
+                            </page>
+                        </notebook>
+                    </sheet>
                 </form>
             </field>
         </record>
@@ -29,13 +41,12 @@
         <record id="view_medical_operational_sector_search" model="ir.ui.view">
             <field name="name">medical.operational.sector.search</field>
             <field name="model">medical.operational.sector</field>
-            <field name="priority" eval="8" />
             <field name="arch" type="xml">
-                <search string="Oemedical Operational_Sector">
-                    <field name="notes" />
-                    <field name="area_id" />
+                <search string="Operational Sectors">
                     <field name="name" />
-                    <newline />
+                    <field name="area_id" />
+                    <field name="notes" />
+
                     <group expand="0" string="Group By...">
                         <filter string="Operational Area" domain="[]"
                             context="{'group_by':'area_id'}" />
@@ -44,33 +55,15 @@
             </field>
         </record>
 
-        <record model="ir.actions.act_window" id="act_open_medical_operational_sector_view">
-            <field name="name">Operational Sector</field>
+        <record id="action_medical_operational_sector_view" model="ir.actions.act_window">
+            <field name="name">Operational Sectors</field>
             <field name="type">ir.actions.act_window</field>
             <field name="res_model">medical.operational.sector</field>
             <field name="view_type">form</field>
             <field name="view_mode">tree,form</field>
-            <field name="search_view_id" ref="view_medical_operational_sector_search" />
-            <field name="domain">[]</field>
-            <field name="context">{}</field>
-        </record>
-
-        <record model="ir.actions.act_window.view" id="act_open_medical_operational_sector_view_form">
-            <field name="act_window_id" ref="act_open_medical_operational_sector_view" />
-            <field name="sequence" eval="20" />
-            <field name="view_mode">form</field>
-            <field name="view_id" ref="view_medical_operational_sector_form" />
-        </record>
-
-        <record model="ir.actions.act_window.view" id="act_open_medical_operational_sector_view_tree">
-            <field name="act_window_id" ref="act_open_medical_operational_sector_view" />
-            <field name="sequence" eval="10" />
-            <field name="view_mode">tree</field>
             <field name="view_id" ref="view_medical_operational_sector_tree" />
+            <field name="search_view_id" ref="view_medical_operational_sector_search" />
         </record>
-
-        <menuitem id="medical_operational_sector" parent="medical.medical_conf_physicians"
-            sequence="70" action="act_open_medical_operational_sector_view" />
 
     </data>
 </openerp>
